# go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Sports analytics API serving player/team profiles, stats, percentiles, news, and journalist tweets. All data-heavy responses are JSON-passthrough from Postgres functions.",
        "title": "Scoracle Data API",
        "contact": {
            "name": "Scoracle"
        },
        "license": {
            "name": "MIT"
        },
        "version": "2.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/news/{entityType}/{entityID}": {
            "get": {
                "description": "Returns news articles about a player or team from Google News RSS (primary) or NewsAPI (fallback). Supports time-window escalation and strict name matching.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get entity news",
                "parameters": [
                    {
                        "enum": [
                            "player",
                            "team"
                        ],
                        "type": "string",
                        "description": "Entity type",
                        "name": "entityType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entityID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "NBA",
                            "NFL",
                            "FOOTBALL"
                        ],
                        "type": "string",
                        "description": "Sport identifier",
                        "name": "sport",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Team name for player context",
                        "name": "team",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max articles (1-50, default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "rss",
                            "api",
                            "both"
                        ],
                        "type": "string",
                        "default": "rss",
                        "description": "News source preference",
                        "name": "source",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile/{entityType}/{entityID}": {
            "get": {
                "description": "Returns a complete player or team profile with bio, current team, and summary stats. Response is raw JSON from Postgres api_player_profile/api_team_profile functions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Get entity profile",
                "parameters": [
                    {
                        "enum": [
                            "player",
                            "team"
                        ],
                        "type": "string",
                        "description": "Entity type",
                        "name": "entityType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entityID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "NBA",
                            "NFL",
                            "FOOTBALL"
                        ],
                        "type": "string",
                        "description": "Sport identifier",
                        "name": "sport",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "detail": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}